+++ pulsar/apps/test/__init__.py
@@ -417,6 +417,14 @@ class TestShowLeaks(TestOption):
     """
 
 
+class TestPep8(TestOption):
+    name = "pep8"
+    flags = ['--pep8']
+    nargs = '*'
+    validator = pulsar.validate_list
+    desc = """Run pep8"""
+
+
 class TestSuite(tasks.TaskQueue):
     '''An asynchronous test suite which works like a task queue.
 
@@ -505,6 +513,12 @@ class TestSuite(tasks.TaskQueue):
             for tag in sorted(_tags()):
                 stream.writeln(tag)
             stream.writeln('')
+            return False
+        elif self.cfg.pep8:
+            msg, code = pep8_run(self.cfg.pep8)
+            stream.writeln(msg)
+            if code:
+                sys.exit(code)
             return False
 
     def monitor_start(self, monitor):