+++ tests/apps/greenio.py
@@ -1,13 +1,30 @@
 import unittest
+from asyncio import Future
 
 try:
     from pulsar.apps import greenio
 except ImportError:
     greenio = None
 
-from pulsar import send, multi_async
+from pulsar import send, multi_async, get_event_loop
 
 from examples.echo.manage import server, Echo
+
+
+def raise_error():
+    raise RuntimeError
+
+
+class EchoGreen(Echo):
+
+    @greenio.run_in_greenlet
+    def __call__(self, message):
+        connection = greenio.wait(self.pool.connect())
+        with connection:
+            consumer = connection.current_consumer()
+            consumer.start(message)
+            result = greenio.wait(consumer.on_finished)
+            return consumer if self.full_response else consumer.buffer
 
 
 @unittest.skipUnless(greenio, 'Requires the greenlet package')
@@ -17,20 +34,39 @@ class TestGreenIO(unittest.TestCase):
     def setUpClass(cls):
         s = server(name=cls.__name__.lower(), bind='127.0.0.1:0')
         cls.server_cfg = yield send('arbiter', 'run', s)
-        cls.client = Echo(cls.server_cfg.addresses[0],
-                          loop=greenio.GreenEventLoop())
+        cls.client = EchoGreen(cls.server_cfg.addresses[0])
 
     @classmethod
     def tearDownClass(cls):
         if cls.server_cfg:
             return send('arbiter', 'kill_actor', cls.server_cfg.name)
 
-    def test_loop(self):
-        c = self.client
-        self.assertIsInstance(c._loop, greenio.GreenEventLoop)
-        self.assertTrue(c._loop._loop)
+    def test_local(self):
+        data = greenio.local()
+        data.bla = 56
+        self.assertEqual(data.bla, 56)
+        def _test_bla():
+            self.assertFalse(hasattr(data, 'bla'))
+            data.bla = 89
+            self.assertEqual(data.bla, 89)
+        gr = greenio.PulsarGreenlet(_test_bla)
+        gr.switch()
+        self.assertEqual(data.bla, 56)
 
-    def test_ping(self):
+    def test_pool(self):
+        pool = greenio.GreenPool()
+        self.assertTrue(pool._loop)
+        self.assertFalse(pool._greenlets)
+        future = pool.submit(lambda: 'Hi!')
+        self.assertIsInstance(future, Future)
+        result = yield future
+        self.assertEqual(result, 'Hi!')
+        # Test an error
+        yield self.async.assertRaises(RuntimeError, pool.submit, raise_error)
+        self.assertEqual(len(pool._greenlets), 2)
+        self.assertEqual(len(pool._available), 2)
+
+    def test_echo(self):
         result = yield self.client(b'ciao luca')
         self.assertEqual(result, b'ciao luca')